Move error helpers from src/log to src/core/error
