+++ src/log.js
@@ -18,6 +18,7 @@ import {
   USER_ERROR_SENTINEL,
   elementStringOrPassThru,
 } from './core/error-message-helpers';
+import {createErrorVargs, duplicateErrorIfNecessary} from './core/error';
 import {findIndex, isArray} from './core/types/array';
 import {getMode} from './mode';
 import {internalRuntimeVersion} from './internal-version';
@@ -621,70 +622,6 @@ export class Log {
       this.assert(assertion, `${opt_message || defaultMessage}: %s`, subject);
     }
   }
-}
-
-/**
- * Some exceptions (DOMException, namely) have read-only message.
- * @param {!Error} error
- * @return {!Error};
- */
-export function duplicateErrorIfNecessary(error) {
-  const messageProperty = Object.getOwnPropertyDescriptor(error, 'message');
-  if (messageProperty && messageProperty.writable) {
-    return error;
-  }
-
-  const {message, stack} = error;
-  const e = new Error(message);
-  // Copy all the extraneous things we attach.
-  for (const prop in error) {
-    e[prop] = error[prop];
-  }
-  // Ensure these are copied.
-  e.stack = stack;
-  return e;
-}
-
-/**
- * @param {...*} var_args
- * @return {!Error}
- * @visibleForTesting
- */
-export function createErrorVargs(var_args) {
-  let error = null;
-  let message = '';
-  for (let i = 0; i < arguments.length; i++) {
-    const arg = arguments[i];
-    if (arg instanceof Error && !error) {
-      error = duplicateErrorIfNecessary(arg);
-    } else {
-      if (message) {
-        message += ' ';
-      }
-      message += arg;
-    }
-  }
-
-  if (!error) {
-    error = new Error(message);
-  } else if (message) {
-    error.message = message + ': ' + error.message;
-  }
-  return error;
-}
-
-/**
- * Rethrows the error without terminating the current context. This preserves
- * whether the original error designation is a user error or a dev error.
- * @param {...*} var_args
- */
-export function rethrowAsync(var_args) {
-  const error = createErrorVargs.apply(null, arguments);
-  setTimeout(() => {
-    // reportError is installed globally per window in the entry point.
-    self.__AMP_REPORT_ERROR(error);
-    throw error;
-  });
 }
 
 /**
